# docs/api/swagger.yaml
definitions:
  github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  github_com_Jozzo6_casino_loyalty_reward_system_internal_types.Promotion:
    properties:
      amount:
        type: number
      created:
        type: string
      description:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      title:
        type: string
      type:
        $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.PromotionType'
      updated:
        type: string
    type: object
  github_com_Jozzo6_casino_loyalty_reward_system_internal_types.PromotionType:
    enum:
    - regular
    - welcome_bonus
    type: string
    x-enum-varnames:
    - Regular
    - WelcomeBonus
  github_com_Jozzo6_casino_loyalty_reward_system_internal_types.TransactionType:
    enum:
    - remove
    - add
    type: string
    x-enum-varnames:
    - TransactionTypeRemove
    - TransactionTypeAdd
  github_com_Jozzo6_casino_loyalty_reward_system_internal_types.User:
    properties:
      balance:
        type: number
      created:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      promotions:
        items:
          $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.UserPromotion'
        type: array
      role:
        $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.UserType'
      updated:
        type: string
    type: object
  github_com_Jozzo6_casino_loyalty_reward_system_internal_types.UserPromotion:
    properties:
      claimed:
        type: string
      created:
        type: string
      end_date:
        type: string
      id:
        type: string
      promotion:
        $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.Promotion'
      promotion_id:
        type: string
      start_date:
        type: string
      updated:
        type: string
      user:
        $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.User'
      user_id:
        type: string
    type: object
  github_com_Jozzo6_casino_loyalty_reward_system_internal_types.UserType:
    enum:
    - 0
    - 1
    type: integer
    x-enum-varnames:
    - Player
    - Staff
  internal_http_users_handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  internal_http_users_handlers.LoginResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      token:
        type: string
    type: object
  internal_http_users_handlers.RegisterRequest:
    properties:
      email:
        type: string
      name:
        minLength: 3
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  internal_http_users_handlers.RegisterResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      token:
        type: string
    type: object
  internal_http_users_handlers.UpdateBalanceRequest:
    properties:
      transaction_type:
        $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.TransactionType'
      value:
        type: number
    required:
    - transaction_type
    - value
    type: object
info:
  contact: {}
paths:
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns their details along with a token.
      parameters:
      - description: User login details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_http_users_handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User logged in successfully
          schema:
            $ref: '#/definitions/internal_http_users_handlers.LoginResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
      summary: Login a user
      tags:
      - Users
...
  /api/v1/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account and returns the user details along with
        a token.
      parameters:
      - description: User registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_http_users_handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User registered successfully
          schema:
            $ref: '#/definitions/internal_http_users_handlers.RegisterResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
      summary: Register a new user
      tags:
      - Users
...
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all users.
      produces:
      - application/json
      responses:
        "200":
          description: List of users retrieved successfully
          schema:
            items:
              $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.User'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
      summary: Get all users
      tags:
      - Users
  /api/v1/users/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a user by their unique ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            type: string
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
      summary: Delete a user
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Retrieves the details of a user by their unique ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User details retrieved successfully
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.User'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
      summary: Get a user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Updates the details of an existing user.
      parameters:
      - description: User details to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.User'
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.User'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
      summary: Update a user
      tags:
      - Users
  /api/v1/users/{id}/balance:
    put:
      consumes:
      - application/json
      description: Updates the balance of a user based on the transaction type and
        value.
      parameters:
      - description: Balance update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_http_users_handlers.UpdateBalanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User balance updated successfully
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.User'
        "400":
          description: Invalid request payload or insufficient balance
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Jozzo6_casino_loyalty_reward_system_internal_types.ErrorResponse'
      summary: Update user balance
      tags:
      - Users
swagger: "2.0"